refactor(backend): collapse DCC models into v0_7_0/dcc.py
Replace the dcc/ package and digital_conformity_credential module with a
single dcc.py module. Update the generator output path and imports.

Made-with: Cursor
