+++ backend/app/models/untp/v0_7_0/__init__.py
@@ -1,6 +1,6 @@
 """UNTP specification v0.7.0 — DCC Pydantic models generated from bundled JSON Schema."""
 
-from app.models.untp.v0_7_0.dcc.digital_conformity_credential import (
+from app.models.untp.v0_7_0.dcc import (
     ConformityAttestation,
     DigitalConformityCredential,
 )
