--- backend/app/models/untp/v0_7_0/dcc/__init__.py
@@ -1,8 +1,0 @@
-"""Digital Conformity Credential (DCC) models for UNTP v0.7.0 (single generated module)."""
-
-from app.models.untp.v0_7_0.dcc.digital_conformity_credential import (
-    ConformityAttestation,
-    DigitalConformityCredential,
-)
-
-__all__ = ["ConformityAttestation", "DigitalConformityCredential"]
